requirements.txt
scikit-learn
label-studio
spacy
mlflow
apache-airflow